feat(Jenkinsfile): separate SonarQube analysis and quality gate stages #main
This commit separates the SonarQube analysis from the build stage, and adds a
new stage for checking the quality gate. If the quality gate status is not OK,
the pipeline will be aborted. This enhances the readability and maintainability
of our Jenkinsfile, and ensures that no code that fails the quality gate is
pushed to the registry.