feat(Jenkinsfile): add git repository name to SonarQube analysis stage
In the Jenkinsfile, an environment variable 'GIT_REPO_NAME' was introduced to store the repository name. A new stage "Initialize Properties" was added to extract the repository name from the 'GIT_URL' and store it in 'GIT_REPO_NAME'. This repository name is then used in SonarQube analysis stage to set 'sonar.projectKey' and 'sonar.projectName'. This enhancement provides more specific context for SonarQube scans, making it easier to track and manage scan results per project.