chore(Jenkinsfile): move nodejs tool declaration to stage, allow any agent #2984303
This commit modifies the Jenkinsfile to allow any agent to run the pipeline,
previously it was restricted to the 'kaniko' label. The usage of nodejs tool
is moved to the "Initialize Properties" stage from the global configuration.
This may simplify the debugging process and improve the flexibility of the
pipeline configuration.